requirements.txt
Pillow==9.0.1
boto3==1.24.70
sanic==22.9.1
ydb==2.10.0
requests==2.28.1
telegram
ydb
six
python-telegram-bot
